requirements.txt
# Core RAG Dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.6

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.10.3

# Vector Store
faiss-cpu>=1.11,<1.13
chromadb==0.4.22

# Google Gemini
google-generativeai==0.3.2

# Embeddings (local alternative)
sentence-transformers==2.3.1

# API Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
tiktoken==0.5.2
numpy==1.26.3
pandas==2.2.0

# Optional: Streamlit for UI
streamlit==1.30.0